% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{sotu_dir}
\alias{sotu_dir}
\title{Save State of the Union Addresses in Directory}
\usage{
sotu_dir(dir = tempfile())
}
\arguments{
\item{dir}{an optional directory of where
the files should be saved. Otherwise,
a temporary location is used. Will be
created if it does not exist. Note
that these temporary files are erased
(eventually) after exiting R.}
}
\value{
a character vector with complete paths
               to the texts
}
\description{
This function saves the State of the Union
Addresses to disk, with one address in each
file. Useful for testing and demonstrating
text analysis functions that read from files.
}
\examples{

fp <- sotu_dir()
txt <- readLines(fp[200])
txt

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
