% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sotu_meta}
\alias{sotu_meta}
\title{Metadata from State of the Union Addresses}
\description{
A data frame containing one row per State of the Union
Address. These line up with the elements in
\code{\link{sotu_text}}.
}
\references{
\url{http://www.presidency.ucsb.edu/sou.php}
}
\keyword{data}
