% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sotu_text}
\alias{sotu_text}
\title{State of the Union Address Text}
\description{
A character vector with one State of the Union in each
element. These line up with the rows of the data in
\code{\link{sotu_meta}}.
}
\references{
\url{http://www.presidency.ucsb.edu/sou.php}
}
\keyword{data}
